Add tests; kill mutants
